OOP data visualizer
